+++ src/main/codegen/util.ts
@@ -4,15 +4,10 @@ export function groupBy<T extends { [key: string]: any }, U extends keyof T>(
 ): [string, T[]][] {
   const groupedItems: { [key: string]: T[] } = {}
 
-  data.forEach(item => {
+  data.forEach((item) => {
     groupedItems[item[key]] = groupedItems[item[key]] || []
     groupedItems[item[key]].push(item)
   })
 
   return Object.entries(groupedItems)
-}
-
-export function intersection<T>(a: T[], b: T[]): T[] {
-  const set = new Set(a)
-  return b.filter(_ => set.has(_))
 }
