+++ src/main/dynamo/GSI.ts
