+++ src/main/dynamo/Model.ts
