+++ src/main/dynamo/Table.ts
