+++ src/main/dynamo/types.ts
