+++ src/main/index.ts
@@ -1,2 +1,2 @@
 export { Beyonce } from "./dynamo/Beyonce"
-export { Table } from "./experimental/Table"
+export { Table } from "./dynamo/Table"
